+++ service/src/main/kotlin/io/provenance/explorer/domain/extensions/DbExtensions.kt
@@ -1,5 +1,6 @@
 package io.provenance.explorer.domain.extensions
 
+import io.provenance.explorer.OBJECT_MAPPER
 import org.jetbrains.exposed.sql.IColumnType
 import org.jetbrains.exposed.sql.transactions.TransactionManager
 import org.joda.time.DateTime
@@ -15,6 +16,16 @@ fun String.exec(args: Iterable<Pair<IColumnType, Any?>>): ResultSet =
         this.fillParameters(args)
         this.executeQuery()
     }
+
+fun <T : Any> String.execAndMap(args: Iterable<Pair<IColumnType, Any?>> = emptyList(), transform: (ResultSet) -> T): List<T> {
+    val result = arrayListOf<T>()
+    TransactionManager.current().exec(this, args) { rs ->
+        while (rs.next()) {
+            result += transform(rs)
+        }
+    }
+    return result
+}
 
 fun <R> ResultSet?.map(transform: (ResultSet) -> R): ArrayList<R> {
     val result = arrayListOf<R>()
@@ -44,3 +55,5 @@ inline fun <T : AutoCloseable?, R> T.use(block: (T) -> R): R {
         }
     }
 }
+
+fun <T> String.mapper(clazz: Class<T>) = OBJECT_MAPPER.readValue(this, clazz)